FunctionApps/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-file-datalake==12.1.2
pyodbc==4.0.30
zeep==4.0.0